refactor(sitemap): merge home + parks into single primary sitemap.xml
sitemap-parks.ts dissolved back into sitemap.ts — home and park pages
are the most important assets and belong in the canonical sitemap.xml.
robots.ts updated to reference 3 sitemaps (sitemap, geo, attractions).
