+++ app/robots.ts
@@ -15,7 +15,6 @@ export default function robots(): MetadataRoute.Robots {
     },
     sitemap: [
       `${baseUrl}/sitemap.xml`,
-      `${baseUrl}/sitemap-parks.xml`,
       `${baseUrl}/sitemap-geo.xml`,
       `${baseUrl}/sitemap-attractions.xml`,
     ],
